Refactor into single task class
